refactor: split cliconfig to own file refactor: normalize CLI caps naming
BREAKING CHANGE: export and name of CLIConfig changed
