+++ src/index.ts
@@ -6,7 +6,8 @@ export { default as SubCommand, UsageExample } from './api/SubCommand';
 export { default as GlobalCommand } from './api/GlobalCommand';
 export { default as GlobalModifierCommand } from './api/GlobalModifierCommand';
 export { default as GroupCommand } from './api/GroupCommand';
-export { default as Context, CliConfig } from './api/Context';
+export { default as Context } from './api/Context';
+export { default as CLIConfig, PluginManagerConfig } from './api/CLIConfig';
 export { default as ServiceFactory } from './api/ServiceFactory';
 export { default as ServiceRegistry } from './api/ServiceRegistry';
 export { default as Service } from './api/Service';
